Rewrite the while loop in DrawingBuffer::createOrRecycleBitmap
Right now inside while loop we call Vector::remove() and that requires
resizing and copying which is quite expensive. This CL rewrite that
for loop by creating a new local Vector, and copy into it all the elements
that are not supposed to be removed. Then we swap this local Vector
with m_recycledBitmaps.

Review-Url: https://codereview.chromium.org/2381613003
Cr-Original-Commit-Position: refs/heads/master@{#421884}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 303824133c0cfff4028742fefe453df5bd222ec5