Fix java_group when depending on testonly targets.
Small fix for when java_groups have testonly targets. Without this
change, I was getting errors such as..

//third_party/hamcrest:hamcrest_java__build_config
which is NOT marked testonly can't depend on
//third_party/hamcrest:hamcrest_core_java__build_config
which is marked testonly.

Also changed some of the Robolectric deps to now be a java
group.

Review-Url: https://codereview.chromium.org/2378713003
Cr-Original-Commit-Position: refs/heads/master@{#421966}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 85331bdd304da37b1d4de21a68d96b42a91d0095