Changed ECaptionSide to an enum class and renamed its members to keywords
Changed ECaptionSide to an enum class and gave it an unsigned underlying
type so it can be stored in a bitfield in the future. Also renamed its
members to match its keywords from CSSValueKeywords.in. This is pre-work
to move ECaptionSide to be generated in ComputedStyleBase.

BUG=628043

Review-Url: https://codereview.chromium.org/2366263002
Cr-Original-Commit-Position: refs/heads/master@{#422029}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fdc8d0ce1cd52ee647079b8a71d6871a87cf1fdc