Fix issue with descendants of aria-activedescendant being marked as focusable.
The ARIA spec says that descendants of a node with aria-activedescendant set should
be marked as focusable if they have an ARIA role. But this doesn't make sense if the
ARIA role is "none" or "presentation", so fix that. Also ensure that we recompute
cached attributes whenever aria-activedescendant changes, even if it doesn't point to
a valid ID, because it can affect focusability.

BUG=651127

Review-Url: https://codereview.chromium.org/2386473003
Cr-Original-Commit-Position: refs/heads/master@{#422164}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6940f7313fe70eb077836a6bdaa8502ffb256ad1