Do not call didAddClient() in finishPendingClients() during revalidation
Revalidation might be started between addClient() and finishPendingClients().
This CL skips didAddClient() in finishPendingClients() if revalidation is
ongoing. In such cases, RawResourceClient::responseReceived() and
dataReceived() are not called in finishPendingClients(), and they are called
later when response headers and payloads for the revalidation is received.

BUG=651763

Review-Url: https://codereview.chromium.org/2382163002
Cr-Original-Commit-Position: refs/heads/master@{#422349}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6faa15bc8dfc6fc9f25c0036b3dc2939f0b1da5e