Make DragCaretController store PositionWithAffinity instead of VisiblePosition
DragCaretController is an on-stack object that stores a VisiblePosition and may
inspect properties of the VisiblePosition even after it is invalidated. This
patch fixes it by making DragCaretController store a PositionWithAffinity
instead.

This patch is a preparation of pruning createVisibleSelectionDeprecated from
DragController: https://codereview.chromium.org/2386083002

BUG=648949,651373

Review-Url: https://codereview.chromium.org/2388703002
Cr-Original-Commit-Position: refs/heads/master@{#422395}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0895689a9cf3613ea07cd1a05ce4c6d32d6d411c