Do not revalidate when ResourceLoaderOptions.synchronousPolicy is different
This CL adds a check of synchronousPolicy to
ResourceLoaderOptions::canReuseRequest() to avoid Resources to be reused
between sync and async requests.
This CL also removes a condition from
ResourceFetcher::determineRevalidationPolicy() that is covered by the newly
added check.

CachingCorrectnessTest is updated so that it doesn't use synchronous
RawResource requests.

BUG=652172

Review-Url: https://codereview.chromium.org/2391523002
Cr-Original-Commit-Position: refs/heads/master@{#422725}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 733a476495c43c1f472e5770c9d8aeeb94c5caca