Fix webkit-patch --help.
I believe that after http://crrev.com/2276713002, invoking webkit-patch
with --help rather than help was broken.

When `webkit-patch help` is invoked, the codepath that's used goes through
the execute method of HelpCommand, and the _tool attribute is set, so it
worked.

But when --help is passed, a different codepath is used, that goes through
optparse which directly uses HelpCommand.help_epilog and the _tool attribute
was never set.

This CL fixes this case by setting the _tool attribute in the constructor
of HelpCommand.

Review-Url: https://codereview.chromium.org/2385373003
Cr-Original-Commit-Position: refs/heads/master@{#423017}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3d1a19a2bd38572c918334f68d83b36989ea557b