Changed EListStylePosition to an enum class and renamed its members to keywords
Changed EListStylePosition to an enum class and gave it an unsigned underlying
type so it can be stored in a bitfield in the future. Also renamed its
members to match its keywords from CSSValueKeywords.in. This is pre-work
to move EListStylePosition to be generated in ComputedStyleBase.

BUG=628043

Review-Url: https://codereview.chromium.org/2380233002
Cr-Original-Commit-Position: refs/heads/master@{#423038}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 063379e6b1aa01de3f305c5343925d9d926e8cd0