Worker: Cancel script loading when an execution context is destroyed
Before this CL, WorkerScriptLoader can work even after the context is destroyed
and a finish callback may crash because of null context access.

This CL cancels script loading when the context is destroyed and provides a way
to check whether the loading is canceled for the finish callback.

BUG=650131

Review-Url: https://codereview.chromium.org/2393783002
Cr-Original-Commit-Position: refs/heads/master@{#423439}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c0837376642c8ef804483ab679d0d7040d844346