PlzNavigate: make ServiceWorker ForeignFetch work
This CL fixes an issue that was preventing foreign fetch to work on
subframes when browser-side-navigation was enabled. The issue was that
there was no way to retrieve the ServiceWorkerContext from the
navigation URLRequest, and so the ServiceWorker specified by the link
header could not be registered.

BUG=440463

Review-Url: https://codereview.chromium.org/2399063002
Cr-Original-Commit-Position: refs/heads/master@{#423833}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 37e10ddfe88bcf59a9f8065b16891ad0ae191f35