[Origin Trials] Clarify the distinction between trials and features
There is some confusion in the origin trial method names and comments
between the name of a feature, which is declared in
RuntimeEnabledFeatures, and a trial, which can enable multiple features.

This CL attempts to make a distinction between the two, most visibly by
renaming OriginTrialContext::isFeatureEnabled to
OriginTrialContext::isTrialEnabled.

R=chasej@chromium.org

Review-Url: https://codereview.chromium.org/2398793003
Cr-Original-Commit-Position: refs/heads/master@{#424033}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a420fe39c98570d80347675134d3738af20c63b7