Don't track SCM changes in rebaseline commands.
There's a fair amount of logic in the rebaseline-related command
which is all for the purpose of tracking which files to add and remove
to the source control management system.

However, since we're now only using git, one possible simpler route
is to just call `git add --all`.

Pros:
 - Much simpler, fewer chances for bugs in SCM change tracking logic.
Cons:
 - This would make rebaseline commands stage all new files from the user
   in the working tree even if they didn't intend to stage these files.
   One possible way to compensate for this would be to *not* call
   `git add --all` after all rebaselining, but instead call this seperately
   when (1) doing a w3c test auto-update or (2) doing a auto-rebaseline
   on rebaseline-o-matic.

This is an alternative to http://crrev.com/2396433004.

BUG=639410

Review-Url: https://codereview.chromium.org/2397573002
Cr-Original-Commit-Position: refs/heads/master@{#424082}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ccd977fa5650bf395c59bbbb0afc81f5dbd3531c