Adding feature usage count for OffscreenCanvas
This CL adds [MeasureAs=xxx] in the .idl files that corresponding to the
OffscreenCanvas creation.

Also, we found that PS#2 has compilation error, and looks like
putting MeasureAs in an interface that has a partial interface
triggers the error. PS#3 is one way to fix it.

BUG=653599, 653913

Review-Url: https://codereview.chromium.org/2398703004
Cr-Original-Commit-Position: refs/heads/master@{#424363}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 99de2c286ec4ad051d562c6e1b7c2955fa860430