[css-flexbox] Don't force layout for overflow: auto anymore
Now that we include the scrollbar in the preferred width due to these changes:
https://chromium.googlesource.com/chromium/src/+/c9b3e99e2a71f7bcc4d9652916347782ff616e0d
https://chromium.googlesource.com/chromium/src/+/906252430fe0108cf163f1fd9c2e96aa283f9cb5%5E%21/

we no longer need to force layout to include the scrollbar in the flex item
size as the above code makes this work correctly. This should help a bit with
performance, too.

(The listed bug is somewhat optimistic; it may help but I haven't really
investigated yet)

R=eae@chromium.org,dgrogan@chromium.org
TEST=n/a, no behavior change
BUG=649988

Review-Url: https://codereview.chromium.org/2411943002
Cr-Original-Commit-Position: refs/heads/master@{#424632}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c03addfdbc3e63d0df45e83f14b6c8d12572ef4b