Clear scroll anchor on all parent scrollers from ScrollAnchor::clear
Before this CL, a call to ScrollAnchor::clear() would clear the anchor node
on the containing scroller, and set the LayoutObject::IsScrollAnchorObject
bit on the anchor if there is no other scroll referencing it.

With this CL, a call to ScrollAnchor::clear() walks up the parent chain and
clears the scroll anchor for every scroller encountered in the walk. As a
result, the LayoutObject::IsScrollAnchorObject is always set to false.

This CL is dependent on https://codereview.chromium.org/2400723002

BUG=650017

Review-Url: https://codereview.chromium.org/2394053004
Cr-Original-Commit-Position: refs/heads/master@{#424748}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 121a3d55a69bb82a46ee9e3159a0be1810052368