Remove unnecessary removeShorthandProperty call in setProperty
The removeShorthandProperty call in MutableStylePropertySet::setProperty
is unnecessary since the function is never called with shorthands, so we
remove the call in this patch.

We probably can't add a DCHECK that the property is not a shorthand as
the class is used for at-rules as well as style rules. A shorthand
property might share the name as a descriptor (I think this is the case
for font-variant), but isShorthandProperty is only for properties.

Review-Url: https://codereview.chromium.org/2415903002
Cr-Original-Commit-Position: refs/heads/master@{#424975}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8fcac47ca0f8061d31321d4cd57ebdfc983277ca