[css-grid] Recompute track sizes if fit-content() argument is modified
When fit-content() was implemented a new attribute
m_fitContentTrackBreadth was added to GridTrackSize.
However we forgot to include it in the == operator implementation,
so we were missing the changes on that value.

Now if you change the argument of a fit-content() track,
the track size is properly recomputed.

BUG=654712
TEST=fast/css-grid-layout/grid-change-fit-content-argument.html

Review-Url: https://codereview.chromium.org/2420673002
Cr-Original-Commit-Position: refs/heads/master@{#425122}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 468f3c355c499087911dbb34c8d2ba6d9a014cff